+++ app/Http/Requests/Admin/AssetFolderRequest.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Http\Requests\Admin;
-
-use WebBlocks\Cms\Http\Requests\Admin\MediaFolderRequest;
-
-class AssetFolderRequest extends MediaFolderRequest {}