+++ app/Http/Requests/Admin/AssetUploadRequest.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace App\Http\Requests\Admin;
-
-use WebBlocks\Cms\Http\Requests\Admin\MediaUploadRequest;
-
-class AssetUploadRequest extends MediaUploadRequest {}