+++ app/Models/Concerns/ValidatesBlockTranslationLocale.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Models\Concerns;
-
-trait ValidatesBlockTranslationLocale
-{
-    use \WebBlocks\Cms\Models\Concerns\ValidatesBlockTranslationLocale;
-}